+++ ldap/servers/plugins/replication/cl5_api.c
@@ -2032,7 +2032,7 @@ static int _cl5AppInit (void)
 	{
 		slapi_log_error(SLAPI_LOG_FATAL, repl_plugin_name_cl, 
 				"_cl5AppInit: failed to fetch backend dbenv (%p) and/or "
-				"index page size (%lu)\n", dbEnv, pagesize);
+				"index page size (%lu)\n", dbEnv, (long unsigned int)pagesize);
 		return CL5_DB_ERROR;
 	}
 }
