+++ ldap/servers/plugins/replication/repl5_replica.c
@@ -437,7 +437,7 @@ replica_get_exclusive_access(Replica *r, PRBool *isInc, PRUint64 connid, int opi
 		slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 				"conn=%" NSPRIu64 " op=%d repl=\"%s\": "
 				"Replica in use locking_purl=%s\n",
-				connid, opid,
+				(long long unsigned int)connid, opid,
 				slapi_sdn_get_dn(r->repl_root),
 				r->locking_purl ? r->locking_purl : "unknown");
 		rval = PR_FALSE;
@@ -450,7 +450,7 @@ replica_get_exclusive_access(Replica *r, PRBool *isInc, PRUint64 connid, int opi
 	{
         slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
 						"conn=%" NSPRIu64 " op=%d repl=\"%s\": Acquired replica\n",
-						connid, opid,
+						(long long unsigned int)connid, opid,
 						slapi_sdn_get_dn(r->repl_root));
 		r->repl_state_flags |= REPLICA_IN_USE;
 		if (isInc && *isInc) 
@@ -491,13 +491,13 @@ replica_relinquish_exclusive_access(Replica *r, PRUint64 connid, int opid)
         slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
 					"conn=%" NSPRIu64 " op=%d repl=\"%s\": "
 					"Replica not in use\n",
-					connid, opid,
+					(long long unsigned int)connid, opid,
 					slapi_sdn_get_dn(r->repl_root));
 	} else {
 		slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name, 
 					"conn=%" NSPRIu64 " op=%d repl=\"%s\": "
 					"Released replica held by locking_purl=%s\n",
-					connid, opid,
+					(long long unsigned int)connid, opid,
 					slapi_sdn_get_dn(r->repl_root),
 					r->locking_purl);
